+++ server.js
@@ -6,7 +6,7 @@ const app = express();
 app.use("/src", express.static(path.resolve(__dirname, "src")));
 
 app.get("/*", (req, res) => {
-    res.sendFile(path.resolve(__dirname, "..", "index.html"));
+    res.sendFile(path.resolve(__dirname, "src", "index.html"));
 });
 
 app.listen(process.env.PORT || 3000, () => {
