+++ src/index.html
@@ -11,9 +11,6 @@
   <a href="/signup" data-link>About</a>
   <a href="/login" data-link>Lorem</a>
     <div id="main"></div>
-    <script type="module" src="/src/scripts/router.js"></script>
-    <script src="/src/index.js"></script>
-    <script src="/src/scripts/login-auth.js"></script>
-    <script src="/src/scripts/signup-auth.js"></script>
+    <script type="module" src="/src/index.js"></script>
 </body>
 </html>
