+++ src/pages/SignupPage.js
@@ -29,7 +29,7 @@ export default class extends AbstractPage {
             class="flex items-center justify-center px-8 py-8 sm:px-12 lg:col-span-7 lg:px-16 lg:py-12 xl:col-span-6"
           >
             <div class="max-w-xl lg:max-w-3xl">
-              <a class="block" href="/" data-link>
+              <a class="block" href="/" data-link="">
                 <span class="sr-only">Home</span>
                 <div class="cursor-pointer"><span class="text-4xl text-[var(--primary)]">to</span><span class="text-4xl bg-clip-text text-transparent bg-gradient-to-tr from-[var(--primary-gradient)] to-[var(--secondary-gradient)]">do.</span></div>
               </a>
@@ -215,6 +215,7 @@ export default class extends AbstractPage {
           </main>
         </div>
       </section>
+    <script src="/src/scripts/signup-auth.js"></script>
             `;
     }
 }
